requirements.txt

reflex==0.8.20
supabase
python-dotenv